tests: Move mock code to tests/common dir.
