+++ core/src/servers.rs
@@ -105,9 +105,3 @@ pub async fn create_aggregator_server(
 
     Ok((client, handle, addr))
 }
-
-#[tracing::instrument(ret(level = tracing::Level::TRACE))]
-fn is_test_env() -> bool {
-    // if thread name is not main then it is a test
-    thread::current().name().unwrap_or_default() != "main"
-}
