+++ core/tests/common/mod.rs
@@ -1,5 +1,15 @@
 //! # Common Utilities for Integration Tests
 
+mod common;
+mod database;
 mod deposit;
+mod env;
+mod rpc;
+mod server;
 
+pub use common::*;
+pub use database::*;
 pub use deposit::*;
+pub use env::*;
+pub use rpc::*;
+pub use server::*;
