--- mock/src/lib.rs
@@ -1,9 +1,0 @@
-//! # Mock Interfaces
-//!
-//! This module includes mock interfaces for tests. There are also some common
-//! elements for unit and integration tests.
-
-pub mod common;
-pub mod database;
-pub mod env;
-pub mod rpc;
